pyproject.toml
[project]
name = "crewai-agents"
version = "0.1.0"
description = "CrewAI agents for code analysis"
requires-python = ">=3.10"
dependencies = [
    "crewai[litellm]>=0.86.0",
    "crewai-tools>=0.17.0",
    "litellm>=1.16.0",
    "google-cloud-aiplatform>=1.42.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
